requirements.txt
Django==1.8.1
django-bootstrap3==5.4.0
django-formtools==1.0
django-tenant-schemas==1.5.2
Pillow==2.8.1
psycopg2==2.6
